+++ src/builders/GCSSBuilder.ts
@@ -11,10 +11,20 @@ export class GCSSBuilder extends GBuilder {
 
   async build() {
     this.src().chain(new CSSPlugin());
-    if (!this.buildOptions.minifyOnly) await this.dest();
+
+    // if outfileOnly is not set, it's default value is treated to be true
+    if (!this.conf.outFile || this.buildOptions.outFileOnly == false) {
+      //---before concat
+      if (this.buildOptions.minify || this.buildOptions.minifyOnly)
+        this.pushStream().chain(GPlugin.cssnano).dest().popStream();
+      if (!this.buildOptions.minifyOnly) this.dest();
+    }
+
+    // concat and the next actions
     if (this.conf.outFile) this.chain(GPlugin.concat);
-    if (this.buildOptions.minify || this.buildOptions.minifyOnly) this.chain(GPlugin.cssnano);
-    return this.dest();
+    if (!this.buildOptions.minifyOnly) this.dest();
+    if (this.buildOptions.minify || this.buildOptions.minifyOnly) this.chain(GPlugin.cssnano).dest();
+    return this;
   }
 }